resolve merge conflicts to have both Chat and history style sheet in app component
