+++ components/HistoryList.js
@@ -1,6 +1,8 @@
 import { useState, useEffect } from 'react';
-import { Card, Container, Row, Col, Modal } from 'react-bootstrap';
+import { Card, Container, Row, Col, Modal, Button, InputGroup, FormControl } from 'react-bootstrap';
 import axios from 'axios';
+import { FaSearch } from "react-icons/fa";
+import moment from 'moment';
 
 const getHistory = (userId, histType) => {
   let urlStr;
@@ -20,24 +22,35 @@ const getHistory = (userId, histType) => {
 const HistListEntry = ( {item} ) => {
   return (
     <div className='hist-list-item' id={item.id}>
-      <img src={item.imageUrl} />
-      <div>
+      <img src={item.imageUrl} className='hist-list-item-img'/>
+      <div className='hist-list-item-info'>
         <div>{item.name}</div>
-        <div>{item.status}</div>
+        <div>{(item.status).charAt(0).toUpperCase() + (item.status).slice(1)}</div>
+        <div>{moment(item.createdAt).format("MM/DD/YYYY")}</div>
       </div>
     </div>
   )
 }
 
-const HistoryList = ( {histType, searchTerm} ) => {
+const HistoryList = ( { histType } ) => {
   const userId = 40;
   const [allHistItems, setAllHistItems] = useState(null);
   const [displayedItems, setDisplayedItems] = useState(null);
+  const [searchTerm, setSearchTerm] = useState(null);
+
+  const handleSearch = (e) =>{
+    let searchStr = e.target.value;
+    if (searchStr.trim().length > 3) {
+      setSearchTerm(searchStr);
+    } else {
+      setSearchTerm(null);
+    }
+  }
 
   useEffect(() => {
     getHistory(userId, histType)
     .then(res => {
-      // console.log('what did server return', res.data);
+      console.log('what did server return', res.data);
       setAllHistItems(res.data);
       setDisplayedItems(res.data);
     })
@@ -48,25 +61,38 @@ const HistoryList = ( {histType, searchTerm} ) => {
 
   useEffect(() => {
     if (displayedItems !== null) {
-      setDisplayedItems(displayedItems.filter((item) => {
-        if (searchTerm && searchTerm.trim().length > 3) {
+      if (searchTerm) {
+        setDisplayedItems(displayedItems.filter((item) => {
           return item.name.toLowerCase().includes(searchTerm.toLowerCase().trim());
-        }
-      }))
-    }
-  }, [userId])
-
-  console.log(searchTerm);
-  console.log(displayedItems);
-  return (
-    <div id='hist-list'>
-      {
-        displayedItems &&
-        displayedItems.map(item => {
-          return <HistListEntry item={item} key={item.id} />
-        })
+        }))
+      } else {
+        setDisplayedItems(allHistItems);
       }
-    </div>
+    }
+  }, [userId, searchTerm])
+
+  // console.log(searchTerm);
+  // console.log(displayedItems);
+  return (
+    <>
+      <InputGroup id='hist-search'>
+          <FormControl
+            placeholder="Search my claims..."
+            onChange={handleSearch}
+          />
+          <Button variant="outline-secondary">
+            <FaSearch />
+          </Button>
+        </InputGroup>
+      <div id='hist-list'>
+        {
+          displayedItems &&
+          displayedItems.map(item => {
+            return <HistListEntry item={item} key={item.id} />
+          })
+        }
+      </div>
+    </>
   )
 }
