+++ package.json
@@ -18,6 +18,7 @@
     "express": "^4.17.1",
     "install": "^0.13.0",
     "jsonwebtoken": "^8.5.1",
+    "moment": "^2.29.1",
     "next": "^12.0.7",
     "npm": "^8.3.0",
     "pg": "^8.7.1",
