+++ pages/history/donations.js
@@ -1,15 +1,9 @@
-import { Button, InputGroup, FormControl, Offcanvas, Nav, Navbar } from 'react-bootstrap';
+import { Button, Offcanvas, Nav, Navbar } from 'react-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css'
-import { FaSearch} from "react-icons/fa";
 import {useState} from 'react';
 import HistoryList from '../../components/HistoryList.js';
 
 const Donations = (props) => {
-  let [searching, setSearching] = useState('');
-
-  const handleSearch = (e) =>{
-    setSearching(e.target.value)
-  }
 
   return (
     <>
@@ -31,17 +25,7 @@ const Donations = (props) => {
         </Navbar.Collapse>
       </Navbar>
 
-      <InputGroup>
-        <FormControl
-          placeholder="Search my donations..."
-          value={searching}
-          onChange={handleSearch}
-        />
-        <Button variant="outline-secondary">
-          <FaSearch />
-        </Button>
-      </InputGroup>
-      <HistoryList histType='donations' searchTerm={searching}/>
+      <HistoryList histType='donations'/>
     </>
   )
 }
